refector: `CoinPrice` 에서 가격 조회를 transaction  -> ticker 로 변경
