+++ CryptocurrencyExchange/CryptocurrencyExchange/Sources/Presentations/Component/CoinPrice/CoinPriceEnvironment.swift
@@ -1,0 +1,12 @@
+//
+//  CoinPriceEnvironment.swift
+//  CryptocurrencyExchange
+//
+//  Created by 이경준 on 2022/03/12.
+//
+
+import Foundation
+
+struct CoinPriceEnvironment {
+    let useCase: TransactionListUseCaseProtocol
+}
