+++ CryptocurrencyExchange/CryptocurrencyExchange/Sources/Presentations/Component/CoinPrice/CoinPriceView.swift
@@ -10,94 +10,11 @@ import SwiftUI
 import ComposableArchitecture
 import Kingfisher
 
-struct CoinPriceState: Equatable {
-    let symbol: String
-    var nowPrice: Double?
-}
 
-extension CoinPriceState {
-    func toViewState() -> OrderBookListViewState {
-        let symbols = symbol.components(separatedBy: "_")
-        let coinSymbol = symbols.first ?? ""
-        let currencySymbol = symbols.last ?? ""
-        let currency = Currency(rawValue: currencySymbol.lowercased()) ?? .none
-        let formattedPrice = nowPrice?.format(to: currency) ?? ""
-        let imageURL = URL(
-            string: "https://cryptoicon-api.vercel.app/api/icon/\(coinSymbol.lowercased())"
-        )
-        return OrderBookListViewState(
-            symbolName: coinSymbol,
-            nowPrice: formattedPrice,
-            imageURL: imageURL
-        )
-    }
-}
 
-struct OrderBookListViewState {
-    let symbolName: String
-    var nowPrice: String
-    let imageURL: URL?
-}
 
-enum CoinPriceAction: Equatable {
-    case onAppear
-    case onDisappear
-    case responseTransactionSingle(Result<[BithumbTransactionHistroySingle], OrderBookListError>)
-    case responseTransactionStream(Result<[BithumbTransactionStream], OrderBookListError>)
-}
 
-struct CoinPriceEnvironment {
-    let useCase: TransactionListUseCaseProtocol
-}
 
-let coinPriceReducer = Reducer<
-    CoinPriceState, CoinPriceAction, CoinPriceEnvironment
-> { state, action, environment in
-    
-    struct CoinPriceCancelId: Hashable {}
-    
-    switch action {
-    case .onAppear:
-        return .merge(
-            environment.useCase
-                .getTransactionHistorySinglePublisher(symbol: state.symbol)
-                .receive(on: DispatchQueue.main)
-                .mapError { OrderBookListError.description($0.localizedDescription) }
-                .catchToEffect(CoinPriceAction.responseTransactionSingle)
-            ,
-            environment.useCase
-                .getTransactionStreamPublisher(symbols: [state.symbol])
-                .receive(on: DispatchQueue.main)
-                .mapError { OrderBookListError.description($0.localizedDescription) }
-                .catchToEffect(CoinPriceAction.responseTransactionStream)
-                .cancellable(id: CoinPriceCancelId())
-        )
-        
-    case .onDisappear:
-        return .cancel(id: CoinPriceCancelId())
-        
-    case let .responseTransactionSingle(.success(transactions)):
-        if let nowPrice = transactions.first?.contPrice {
-            state.nowPrice = nowPrice
-        }
-        return .none
-        
-    case let .responseTransactionSingle(.failure(error)):
-        Log.error(error)
-        return .none
-        
-    case let .responseTransactionStream(.success(transactions)):
-        if let nowPrice = transactions.first?.contPrice {
-            state.nowPrice = nowPrice
-        }
-        return .none
-        
-    case let .responseTransactionStream(.failure(error)):
-        Log.error(error)
-        return .none
-        
-    }
-}
 
 struct CoinPriceView: View {
     let store: Store<CoinPriceState, CoinPriceAction>
